move confirm button to app dir
