+++ apps/next/app/confirm-button/head.tsx
@@ -1,0 +1,11 @@
+import { DefaultTags } from '@/ui/DefaultTags'
+
+export default function Head() {
+	return (
+		<>
+			<DefaultTags />
+			<title>Confirm button</title>
+			<meta name="description" content="Confirm button" />
+		</>
+	)
+}
