+++ apps/next/tsconfig.json
@@ -27,6 +27,6 @@
 		],
 		"strictNullChecks": true
 	},
-	"include": ["*.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
+	"include": ["*.d.ts", "**/*.ts", "**/*.tsx"],
 	"exclude": ["node_modules"]
 }
